Dtype performance improvements

+++ backends/src/dtype.rs
@@ -54,7 +54,7 @@ impl Default for DType {
             feature = "ort"
         )))]
         {
-            DType::Float16
+            DType::BFloat16
         }
     }
 }
